update all warehouses

+++ src/main/java/org/slavik/breez/model/Price.java
@@ -3,16 +3,19 @@ package org.slavik.breez.model;
 import com.fasterxml.jackson.annotation.*;
 
 public class Price {
-    private int ric;
+    private double ric;
     private String ricCurrency;
 
+
     @JsonProperty("ric")
-    public int getRic() { return ric; }
+    public int getRic() { return (int) ric; }
     @JsonProperty("ric")
-    public void setRic(int value) { this.ric = value; }
+    public void setRic(double value) { this.ric = value; }
 
     @JsonProperty("ric_currency")
     public String getRicCurrency() { return ricCurrency; }
     @JsonProperty("ric_currency")
     public void setRicCurrency(String value) { this.ricCurrency = value; }
+
+
 }
